Add stress tests for slicer performance
These are intended to produce very big models or with lots of parts so that the slicer performance is put to the test.
I had more models in here but those are pre-generated STLs with file sizes ranging up to 700MB so that's not appropriate to put in Git or upload anywhere.