Some preferential adjustments
The thickness of 4mm was a bit overdone. We could save some material and printing time by reducing this.
We need a bit of extra height in the holder in order to make the motor not hit the ground with the extra height now, though.
Also, increase the radius of the rod gear. According to the internet, we should be aiming for about 13RPM on the tumbler. We're not going to make that by far, with our 20RPM motor and our gear ratio of 7/26, but with a slightly bigger gear we should get much closer.