Remove hole for nut
The nut must now be on the inside of the drum, exposed to the wear and tear of the rocks and such but no longer inside the wall. It turned out that the nut was too thick to fit inside that wall. Or rather, it was just thin enough but the resolution of the printer couldn't deal with that thin a wall, and walls of 0.15mm wouldn't be strong enough anyway.