Add stamp for SiU
I traced the S and the I from the logo, then designed the U in the same style (since it is normally in a different font below). Then I exported it from InkScape into OpenSCAD using this plug-in for InkScape: https://github.com/martymcguire/inkscape-openscad-poly
The rest is just some coding in OpenSCAD to extrude it and place a cylinder on top.
The cylinder needs to be on top (not below) because otherwise it mirrors the stamp. Please manually rotate it in the slicer.