Add more simple model for narrow passage
The other similar models we have have various degrees of narrowing at different heights. And curves to create the narrowing, or narrow until 0. But for variable-width lines we need to make it continue broadening on the other end. And to allow debugging vertex-by-vertex, this model consists of more straight parts.