Inset the game squares for better interlocking
When printing in dual extrusion with the game squares a different colour, the game squares will now be lowered into the body of the game board a bit. This interlocks them with the game board, making them sturdier. It also looks a bit fancier since the relief of the squares is greater without making the game board thicker.