Graduate towards a real heart shape
I haven't really figured out yet how to make it combine the cardioid function with the halfcircle. I kind of need a linear interpolation depending on b, alternating between the cardioid and the halfcircle at every 90 degrees. Shouldn't be too difficult, but it's best to save it here for now.